Add EVENTROUTER to handle usb suspend/resume (hotplugging)
VBUS_CHARGE bit is set to fool the usb peripheral to think there is a
host attached. This will force suspend mode when bus is inactive (when
no host is attached). This in turn will save power when device boots
without host attached, but still enables USB hotplugging when the host
is attached later on.